Add Responsive Suffixes
These suffixes take the format @<breakpoint>, and tell us `this class
when at this size`:

 <ul class="u-hidden@palm">
   <li>...</li>
   <li>...</li>
   <li>...</li>
 </ul>

The @ is a human readable and logical way of denoting conditional
states. It allows developers to learn about any potential permutations
or appearances that the piece of UI in question might have, just at a
glance.